+++ src/Context/WordpressContext.php
@@ -3,7 +3,7 @@ namespace PaulGibbs\WordpressBehatExtension\Context;
 
 use InvalidArgumentException;
 use Behat\Gherkin\Node\TableNode;
-use function PaulGibbs\WordpressBehatExtension\is_wordpress_error;
+use function PaulGibbs\WordpressBehatExtension\Util\is_wordpress_error;
 
 /**
  * Provides step definitions for a range of common tasks. Recommended for all test suites.